+++ package.json
@@ -29,7 +29,6 @@
   "homepage": "https://github.com/Samu-K/pomodo#readme",
   "dependencies": {
     "@date-io/date-fns": "^3.2.1",
-    "@mdi/js": "^7.4.47",
     "@tanstack/vue-query": "^5.90.7",
     "@tauri-apps/api": "^2",
     "@tauri-apps/plugin-dialog": "^2.4.2",
